Move is_subgraph to subgraph module
