+++ strawberryfields/gbs/clique.py
@@ -73,7 +73,7 @@ Code details
 import networkx as nx
 import numpy as np
 
-from strawberryfields.gbs import utils
+from strawberryfields.gbs import subgraph as sub
 
 
 def search(clique: list, graph: nx.Graph, iterations, node_select: str = "uniform") -> list:
@@ -175,7 +175,7 @@ def grow(clique: list, graph: nx.Graph, node_select: str = "uniform") -> list:
         list[int]: a new clique subgraph of equal or larger size than the input
     """
 
-    if not utils.is_subgraph(clique, graph):
+    if not sub.is_subgraph(clique, graph):
         raise ValueError("Input is not a valid subgraph")
 
     if not is_clique(graph.subgraph(clique)):
@@ -235,7 +235,7 @@ def swap(clique: list, graph: nx.Graph, node_select: str = "uniform") -> list:
            list[int]: a new clique subgraph of equal size as the input
        """
 
-    if not utils.is_subgraph(clique, graph):
+    if not sub.is_subgraph(clique, graph):
         raise ValueError("Input is not a valid subgraph")
 
     if not is_clique(graph.subgraph(clique)):
@@ -283,7 +283,7 @@ def shrink(subgraph: list, graph: nx.Graph) -> list:
         list[int]: a clique of size smaller than or equal to the input subgraph
     """
 
-    if not utils.is_subgraph(subgraph, graph):
+    if not sub.is_subgraph(subgraph, graph):
         raise ValueError("Input is not a valid subgraph")
 
     subgraph = graph.subgraph(subgraph).copy()  # A copy is required to be able to modify the
